requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.12
django-crispy-forms==1.13.0
django-filter==21.1
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
psycopg2==2.9
pytz==2021.3
razorpay==1.3.0
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
whitenoise==5.3.0
